requirements.txt
pandas
numpy 
pymongo
certifi
python-dotenv
setuptools
pymongo[srv]==3.11
scikit-learn
dill
pyaml
# -e .